+++ src/grepros/main.py
@@ -8,7 +8,7 @@ Released under the BSD License.
 
 @author      Erki Suurjaak
 @created     23.10.2021
-@modified    08.12.2021
+@modified    10.12.2021
 ------------------------------------------------------------------------------
 """
 ## @namespace grepros.main
@@ -269,7 +269,7 @@ Export all bag messages to SQLite and Postgres, print only export progress:
              default=[], nargs="*", type=lambda x: (x.split("=", 1)*2)[:2],
              help="write options as key=value pairs, supported flags:\n"
                   "  template=/my/path.tpl - custom template to use for HTML output\n"
-                  "  commit-interval=NUM - transaction size for Postgres output\n"
+                  "  commit-interval=NUM - transaction size for Postgres/SQLite output\n"
                   "                        (default 1000, 0 is autocommit)\n"
                   "  nesting=array|all - create tables for nested message types\n"
                   "                      in Postgres/SQLite output,\n"